requirements.txt
Flask==2.3.3
face-recognition==1.3.0
Pillow==10.0.1
Werkzeug==2.3.7
opencv-python==4.8.1.78
dlib==19.24.2
numpy==1.24.3
scikit-image==0.21.0
